requirements.txt
llama-index
faiss-cpu
llama-index-vector-stores-faiss
orjson
numpy
docstring-parser
